Fixed some bugs introduced through previous tweaks that I noticed through difference in reduced chiSq and run time.

+++ PEFinder/Globals.h
@@ -22,9 +22,9 @@ extern float pdfT0SampleConv; // Conversion of sampling rate to a double
 extern double WFSigThresh; // Threshold for considering either PEs or after-pulses (V).
 extern int maxPEs;  // Maximum number of PEs RecoMore will find before stopping looking and doing the fit.
 
-extern float ampDiff; // Running average of the difference between initial amplitude guess and amplitude fit value.
-extern float timeDiff; // Running average of the difference between initial time guess and time fit value.
-extern float baselineDiff; // Running average of the difference between initial baseline guess and baseline fit value.
+extern double ampDiff; // Running average of the difference between initial amplitude guess and amplitude fit value.
+extern double timeDiff; // Running average of the difference between initial time guess and time fit value.
+extern double baselineDiff; // Running average of the difference between initial baseline guess and baseline fit value.
 
 extern int sysProcPECount; // Count of number of PEs processed during entire application run. Used to calculate average error in parameter estimate.
 extern int sysProcWFCount; // Count of number of waveforms processed during entire application run. Used for average reduced chiSq.
